ventas/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
psycopg[binary]>=3.2.6
python-dotenv==1.0.1